+++ packages/plugins/eslint-plugin/src/configs/disable-experimental.ts
@@ -17,7 +17,8 @@ export const rules: Record<string, RuleConfig> = {
   "@eslint-react/unstable-rules-of-props": "off",
   "@eslint-react/unstable-rules-of-state": "off",
 
-  "@eslint-react/jsx-runtime": "off",
+  "@eslint-react/jsx-no-deoptimization": "off",
+  "@eslint-react/jsx-no-namespace": "off",
 
   "@eslint-react/rsc-function-definition": "off",
 };